+++ src/main/java/com/emmkay/infertility_system_api/dto/response/ManagerDashboardStatisticsResponse.java
@@ -9,7 +9,7 @@ import lombok.experimental.FieldDefaults;
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class ManagerDashboardStatisticsResponse {
-    int totalRevenue;
-    int totalAppointments;
-    int totalCustomersTreated;
+    Long totalRevenue;
+    Long totalAppointments;
+    Long totalCustomersTreated;
 }
